updated code with navigations, api calls and ui changes

+++ src/api/permissionsApi.js
@@ -2,24 +2,24 @@ import api from '../config/axiosConfig';
 
 // Get all permissions
 export const getAllPermissions = () =>
-  api.get('/permissions').then(res => res.data);
+  api.get('/admin/permissions').then(res => res.data);
 
 // Get single permission by ID
 export const getPermissionById = (id) =>
-  api.get(`/permissions/${id}`).then(res => res.data);
+  api.get(`/admin/permissions/${id}`).then(res => res.data);
 
 // Get permissions by category
 export const getPermissionsByCategory = (category) =>
-  api.get(`/permissions/category/${category}`).then(res => res.data.permissions);
+  api.get(`/admin/permissions/category/${category}`).then(res => res.data.permissions);
 
 // Create new permission
 export const createPermission = (permissionData) =>
-  api.post('/permissions', permissionData).then(res => res.data.permission || res.data);
+  api.post('/admin/permissions', permissionData).then(res => res.data.permission || res.data);
 
 // Update a permission
 export const updatePermission = (id, data) =>
-  api.put(`/permissions/${id}`, data).then(res => res.data.permission || res.data);
+  api.put(`/admin/permissions/${id}`, data).then(res => res.data.permission || res.data);
 
 // Delete a permission
 export const deletePermission = (id) =>
-  api.delete(`/permissions/${id}`).then(res => res.data);
+  api.delete(`/admin/permissions/${id}`).then(res => res.data);
